+++ demisto_sdk/commands/common/tests/pack_metadata_validator_test.py
@@ -1,12 +1,14 @@
 import io
 import os
+from typing import Dict
 
 import pytest
 from demisto_sdk.commands.common import tools
+from demisto_sdk.commands.common.constants import EXCLUDED_DISPLAY_NAME_WORDS
 from demisto_sdk.commands.common.hook_validations.base_validator import \
     BaseValidator
-from demisto_sdk.commands.common.hook_validations.pack_unique_files import \
-    PackUniqueFilesValidator
+from demisto_sdk.commands.common.hook_validations.pack_unique_files import (
+    PACK_METADATA_NAME, PACK_METADATA_SUPPORT, PackUniqueFilesValidator)
 from demisto_sdk.commands.common.legacy_git_tools import git_path
 
 
@@ -48,6 +50,47 @@ class TestPackMetadataValidator:
 
         validator = PackUniqueFilesValidator('fake')
         assert not validator.validate_pack_meta_file()
+
+    VALIDATE_PACK_NAME_INPUTS = [({PACK_METADATA_NAME: 'fill mandatory field'}, False),
+                                 ({PACK_METADATA_NAME: 'A'}, False),
+                                 ({PACK_METADATA_NAME: 'notCapitalized'}, False),
+                                 ({PACK_METADATA_NAME: 'BitcoinAbuse (Community)', PACK_METADATA_SUPPORT: 'community'},
+                                  False),
+                                 ({PACK_METADATA_NAME: 'BitcoinAbuse'}, True)]
+
+    @pytest.mark.parametrize('metadata_content, expected', VALIDATE_PACK_NAME_INPUTS)
+    def test_validate_pack_name(self, metadata_content: Dict, expected: bool, mocker):
+        """
+        Given:
+        - Metadata JSON pack file content.
+
+        When:
+        - Validating if pack name is valid.
+
+        Then:
+        - Ensure expected result is returned.
+        """
+        validator = PackUniqueFilesValidator('fake')
+        mocker.patch.object(validator, '_add_error', return_value=True)
+        assert validator.validate_pack_name(metadata_content) == expected
+
+    def test_name_does_not_contain_excluded_word(self):
+        """
+        Given:
+        - Pack name.
+
+        When:
+        - Validating pack name does not contain excluded word.
+
+        Then:
+        - Ensure expected result is returned.
+        """
+        pack_name: str = 'Bitcoin Abuse'
+        validator = PackUniqueFilesValidator('fake')
+        assert validator.name_does_not_contain_excluded_word(pack_name)
+        for excluded_word in EXCLUDED_DISPLAY_NAME_WORDS:
+            invalid_pack_name: str = f'{pack_name} ({excluded_word})'
+            assert not validator.name_does_not_contain_excluded_word(invalid_pack_name)
 
     @staticmethod
     def read_file(file_):