+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * The **init** command will now change the script template name in the code to the given script name.
 * Expanded the validations performed on beta integrations.
 * Improved the error messages in **generate-docs**, if an example was not provided.
+* Added to **validate** command a validation that a content entity or a pack name does not contain the words "partner" and "community".
+
 # 1.4.5
 * Enhanced the **postman-codegen** command to name all generated arguments with lower case.
 * Fixed an issue where the **find-dependencies** command miscalculated the dependencies for playbooks that use generic commands.